fix: remove use-scroll-top hook and merged it in use=page-scroll
